requirements.txt
matplotlib==3.3.2
numpy==1.19.2
efficientnet==1.1.1
tensorflow==2.6.0
pandas==1.1.3
scipy==1.3.1
tqdm==4.50.2
tensorflow_addons==0.14.0
scikit_image==0.17.2
scikit_learn==0.24.2
umap-learn>=0.3.10